refactor: move parallel reader to separate module
