+++ src/parallel/mod.rs
@@ -2,8 +2,9 @@ mod error;
 pub(crate) mod multi;
 pub(crate) mod paired;
 mod processor;
+pub(crate) mod reader;
 pub(crate) mod single;
 
 pub use error::{IntoProcessError, ProcessError, Result};
 pub use processor::{MultiParallelProcessor, PairedParallelProcessor, ParallelProcessor};
-pub use single::ParallelReader;
+pub use reader::ParallelReader;
